Connected menus to eachother
