+++ Assets/Prefabs/Menus/Credits Menu.prefab
@@ -400,8 +400,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 7155479639203331531}
-        m_TargetAssemblyTypeName: UIMenus, Assembly-CSharp
+      - m_Target: {fileID: 1598950832950253420}
+        m_TargetAssemblyTypeName: UICreditsMenu, Assembly-CSharp
         m_MethodName: MainMenu
         m_Mode: 1
         m_Arguments:
@@ -503,9 +503,9 @@ GameObject:
   - component: {fileID: 2471339842166900290}
   - component: {fileID: 4900732066448075900}
   - component: {fileID: 1550679749547940128}
-  - component: {fileID: 7155479639203331531}
+  - component: {fileID: 1598950832950253420}
   m_Layer: 5
-  m_Name: Credits
+  m_Name: Credits Menu
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -594,7 +594,7 @@ MonoBehaviour:
   m_BlockingMask:
     serializedVersion: 2
     m_Bits: 4294967295
---- !u!114 &7155479639203331531
+--- !u!114 &1598950832950253420
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -603,7 +603,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 3886916768297354923}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 94659c5fc02ddf44b9fa4ff16ab46b6f, type: 3}
+  m_Script: {fileID: 11500000, guid: afe359c77c2563d40922ee9492c5fda6, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &3888311182921659864
