+++ Assets/Prefabs/Menus/Settings Menu.prefab
@@ -791,7 +791,7 @@ GameObject:
   - component: {fileID: 5823436052750804442}
   - component: {fileID: 6264290467216499452}
   m_Layer: 5
-  m_Name: SettingsMenu
+  m_Name: Settings Menu
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
