+++ Assets/Prefabs/Menus/UICreditsMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: afe359c77c2563d40922ee9492c5fda6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
