+++ Assets/Prefabs/Menus/UILoseMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 769ba5dab50133c488d6a2d0ee3d2924
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
