--- Assets/Prefabs/Menus/UIMenus.cs
@@ -1,42 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class UIMenus : MonoBehaviour
-{
-
-    public void QuitGame()
-    {
-        Application.Quit();
-    }
-
-    public void Startgame()
-    {
-        
-    }
-
-    public void MainMenu()
-    {
-
-    }
-
-    public void Customize()
-    {
-
-    }
-
-    public void Credits()
-    {
-
-    }
-
-    public void Settings()
-    {
-
-    }
-
-    public void ReturnFromSettings()
-    {
-        
-    }
-}
