+++ Assets/Prefabs/Menus/UISettingsMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3ad77f9ccf89a164fbb8ce4c9a603921
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
