+++ Assets/Prefabs/Menus/UIWinMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4218382e08b6b104aa5f73f2b70fe475
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
